PhotoCollectionCell set in StoryBoard
