+++ Photo Collection/Photo Collection/Model + Model Controller/Photo.swift
