+++ Photo Collection/Photo Collection/Model + Model Controller/PhotoController.swift
