+++ Photo Collection/Photo Collection/Other/Assets.xcassets/AppIcon.appiconset/Contents.json
