+++ Photo Collection/Photo Collection/Other/Assets.xcassets/Contents.json
