+++ Photo Collection/Photo Collection/Other/Info.plist
