+++ Photo Collection/Photo Collection/PhotosCollectionViewCell.swift
@@ -1,0 +1,21 @@
+//
+//  PhotosCollectionViewCell.swift
+//  Photo Collection
+//
+//  Created by Rick Wolter on 10/17/19.
+//  Copyright © 2019 Richar Wolter. All rights reserved.
+//
+
+import UIKit
+
+class PhotosCollectionViewCell: UICollectionViewCell {
+    
+    
+    
+    
+    
+    
+    
+    
+    
+}
