--- Photo Collection/Photo Collection/ViewController.swift
@@ -1,20 +1,0 @@
-//
-//  ViewController.swift
-//  Photo Collection
-//
-//  Created by Rick Wolter on 10/17/19.
-//  Copyright © 2019 Richar Wolter. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
